requirements.txt
asgiref==3.8.1
coverage==7.6.11
Django==5.1.5
djangorestframework==3.15.2
drf-material==0.3.3
drf-redesign==0.4.0
iniconfig==2.0.0
isort==5.13.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
PyYAML==6.0.2
sqlparse==0.5.3
